[RENAME] projectSearchMetadataGET.js
